# Remove or split up Frame methods that use the index