--- a/python/cudf/cudf/core/frame.py
+++ b/python/cudf/cudf/core/frame.py
@@ -1285,12 +1285,17 @@ class Frame(BinaryOperand, Scannable):
             else:
                 filled_data[col_name] = col.copy(deep=True)
 
+        # TODO: This logic needs to move into the IndexedFrame class.
+        old_index = self._index
         ret = self._mimic_inplace(
             self._from_data(data=filled_data), inplace=inplace,
         )
         # TODO: Split this logic into the IndexedFrame class.
-        if isinstance(ret, cudf.core.indexed_frame.IndexedFrame):
-            ret._index = self._index
+        if isinstance(self, cudf.core.indexed_frame.IndexedFrame):
+            if inplace:
+                self._index = old_index
+            else:
+                ret._index = old_index
         return ret
 
     @_cudf_nvtx_annotate