--- a/python/cudf/cudf/core/frame.py
+++ b/python/cudf/cudf/core/frame.py
@@ -397,7 +397,7 @@ class Frame(BinaryOperand, Scannable):
         return result
 
     @_cudf_nvtx_annotate
-    def equals(self, other, **kwargs):
+    def equals(self, other):
         """
         Test whether two objects contain the same elements.
         This function allows two Series or DataFrames to be compared against
@@ -454,30 +454,19 @@ class Frame(BinaryOperand, Scannable):
         >>> df.equals(different_column_type)
         True
         """
-        if self is other:
-            return True
-
-        check_types = kwargs.get("check_types", True)
-
-        if check_types:
-            if type(self) is not type(other):
-                return False
-
-        if other is None or len(self) != len(other):
+        if (
+            other is None
+            or not isinstance(other, type(self))
+            or len(self) != len(other)
+        ):
             return False
 
-        # check data:
-        for self_col, other_col in zip(
-            self._data.values(), other._data.values()
-        ):
-            if not self_col.equals(other_col, check_dtypes=check_types):
-                return False
-
-        # check index:
-        if self._index is None:
-            return other._index is None
-        else:
-            return self._index.equals(other._index)
+        return all(
+            self_col.equals(other_col, check_dtypes=True)
+            for self_col, other_col in zip(
+                self._data.values(), other._data.values()
+            )
+        )
 
     @_cudf_nvtx_annotate
     def _get_columns_by_label(self, labels, downcast=False):