--- a/python/cudf/cudf/core/indexed_frame.py
+++ b/python/cudf/cudf/core/indexed_frame.py
@@ -194,7 +194,7 @@ class IndexedFrame(Frame):
         return out
 
     @classmethod
-    @annotate("FRAME_FROM_COLUMNS", color="green", domain="cudf_python")
+    @annotate("INDEXEDFRAME_FROM_COLUMNS", color="green", domain="cudf_python")
     def _from_columns(
         cls,
         columns: List[ColumnBase],
@@ -225,7 +225,9 @@ class IndexedFrame(Frame):
         return out
 
     @annotate(
-        "FRAME_FROM_COLUMNS_LIKE_SELF", color="green", domain="cudf_python"
+        "INDEXEDFRAME_FROM_COLUMNS_LIKE_SELF",
+        color="green",
+        domain="cudf_python",
     )
     def _from_columns_like_self(
         self,