… python/cudf/cudf/core/multiindex.py
… python/cudf/cudf/core/series.py
… python/cudf/cudf/core/single_column_frame.py
… python/cudf/cudf/tests/test_dataframe.py