<!--

Thank you for contributing to cuDF :)

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. There are CI checks in place to enforce that committed code follows our style
   and syntax standards. Please see our contribution guide in `CONTRIBUTING.MD`
   in the project root for more information about the checks we perform and how
   you can run them locally.

4. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

5. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please mark your pull request as Draft.
   https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/changing-the-stage-of-a-pull-request#converting-a-pull-request-to-a-draft

6. If your pull request is ready to be reviewed without requiring additional
   work on top of it, then remove it from "Draft" and make it "Ready for Review".
   https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/changing-the-stage-of-a-pull-request#marking-a-pull-request-as-ready-for-review

   If assistance is required to complete the functionality, for example when the
   C/C++ code of a feature is complete but Python bindings are still required,
   then add the label `help wanted` so that others can triage and assist.
   The additional changes then can be implemented on top of the same PR.
   If the assistance is done by members of the rapidsAI team, then no
   additional actions are required by the creator of the original PR for this,
   otherwise the original author of the PR needs to give permission to the
   person(s) assisting to commit to their personal fork of the project. If that
   doesn't happen then a new PR based on the code of the original PR can be
   opened by the person assisting, which then will be the PR that will be
   merged.

7. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please do not
   rebase your branch on the target branch, force push, or rewrite history.
   Doing any of these causes the context of any comments made by reviewers to be lost.
   If conflicts occur against the target branch they should be resolved by
   merging the target branch into the branch used for making the pull request.

Many thanks in advance for your cooperation!

-->
This PR contributes to excising indexes from Frame and moving them entirely to IndexedFrame. A number of methods (such as `copy`, `equals`, and `_mimic_inplace`) have either been removed from `Frame` or had the index-related logic moved to corresponding function overrides in IndexedFrame. In the process, this PR also implements certain optimizations as part of the rewrite. Of particular interest is that because indexes are immutable, `IndexedFrame.copy` has been modified to _always_ shallow copy the index, which should significantly reduce memory pressure in copy-heavy applications. This also results in significant performance improvements when the Index is not a RangeIndex, for example:
```
In [2]: df = cudf.DataFrame({"a": range(10000), "b": range(10000)}, index=[i*2 for i in range(10000)])
# Before
In [3]: %timeit df.copy(deep=True)
151 µs ± 2.05 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# After
In [3]: %timeit df.copy(deep=True)
119 µs ± 804 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

Other methods rewritten in this PR have also been sped up, although the speedups are typically more modest. For example:
```
# Using the same DataFrame as above
# Before
In [4]: %timeit df.clip(100, 1000)
489 µs ± 4.39 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# After
In [4]: %timeit df.clip(100, 1000)
431 µs ± 2.42 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```